Implement ExampleGenerator using SPI
